feat() Add ChatPage and MainDashboard components for enhanced chat functionality
